TST split numpy tests into two files. added fft type tests
